refactor: 변수 및 파라미터 final 추가
